Fix loop time issue with date time

+++ src/Cells/DateTimeCell.h
@@ -7,7 +7,7 @@ class DateTimeCell : public Cell
         bool _showDate = false;
         bool _showTime = false;
         bool _lastValid = false;   
-        uint8_t _lastYear = 0;
+        uint16_t _lastYear = 0;
         uint8_t _lastMonth = 0;
         uint8_t _lastHour = 0;
         uint8_t _lastMinute = 0;
