+++ src/Pages/DateTimePage.h
@@ -4,7 +4,7 @@
 class DateTimePage : public Page
 {
     bool _lastValid = false;   
-    uint8_t _lastYear = 0;
+    uint16_t _lastYear = 0;
     uint8_t _lastMonth = 0;
     uint8_t _lastHour = 0;
     uint8_t _lastMinute = 0;
